updated class name for links, organized components and moved Browser Router to index.js
