+++ package.json
@@ -45,6 +45,6 @@
     ]
   },
   "devDependencies": {
-    "cypress": "^3.8.1"
+    "cypress": "^3.8.3"
   }
 }
