+++ src/App.css
@@ -20,6 +20,6 @@
 .App-link {
   color: #09d3ac;
 }
-.bold {
+.selected-view {
   font-weight: bold;
 }
